renaming common components package
