+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/home/HomeScreen.kt
@@ -42,7 +42,7 @@ import br.com.rbrthmn.R
 import br.com.rbrthmn.ui.financialcompanion.screens.home.components.BalanceCard
 import br.com.rbrthmn.ui.financialcompanion.screens.home.components.CreditCardBillsCard
 import br.com.rbrthmn.ui.financialcompanion.screens.home.components.LastMonthDifferenceCard
-import br.com.rbrthmn.ui.financialcompanion.components.MonthSelectionTopBar
+import br.com.rbrthmn.ui.financialcompanion.commoncomponents.MonthSelectionTopBar
 import br.com.rbrthmn.ui.financialcompanion.screens.home.components.MonthlyLimitCard
 import br.com.rbrthmn.ui.financialcompanion.navigation.NavigationDestination
 import br.com.rbrthmn.ui.financialcompanion.utils.MonthsOfTheYear
